Add screenshot of merge sort resulting data
This PR adds a new screenshot image `s04-mergeSort-Imperial-Screengrac-Data.png` to the `images` directory. This image shows the resulting data after running the merge sort algorithm as part of the coursework for the Algorithms and Complexity class.

Changes:
- Added `images/s04-mergeSort-Imperial-Screengrac-Data.png`

This image will be used to visually represent the output of the merge sort process in the documentation.