Add screenshot of potential user input for merge sort
This PR adds a new screenshot image `s04-mergeSort-Imperial-Screengrac-Input.png` to the `images` directory. This image shows a potential user input for the merge sort algorithm as part of the coursework for the Algorithms and Complexity class.

Changes:
- Added `images/s04-mergeSort-Imperial-Screengrac-Input.png`

This image will be used to visually represent a sample user input for the merge sort process in the documentation.