Add screenshot of potential script output for merge sort
This PR adds a new screenshot image `s04-mergeSort-Imperial-Screengrac-Output.png` to the `images` directory. This image shows a potential output for the merge sort algorithm as part of the coursework for the Algorithms and Complexity class.

Changes:
- Added `images/s04-mergeSort-Imperial-Screengrac-Output.png`

This image will be used to visually represent a sample script output for the merge sort process in the documentation.